v0_initial_flow_n8n-python/python/requirements.txt
# Python dependencies for materials extraction pipeline
google-generativeai
pypdf2
pdfplumber
python-magic
